Make language suffix matching case-insensitive
This fixes the case where the language folder was created
in lowercase, e.g. 'pt-br', and Smartcat maps it `pt-BR` language,
but then on `pull` it fails to remove the `_pt-br` file name
suffix, generates an incorrect local file name, doesn't find it
and doesn't download the .po files from the server.